FIX: fix add goal button to display on absolute position
