+++ src/pages/Home.tsx
@@ -1,14 +1,13 @@
 import React from 'react';
-import { useNavigate } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
 
 import UserProfile from '../components/user/UserProfile';
 import MyGoalCard from '../components/goal/MyGoalCard';
-import Icon from '../components/common/elem/Icon';
 import Alert from '../components/common/alert/Alert';
 import LoadingMsg from '../components/common/elem/LoadingMsg';
 import ErrorMsg from '../components/common/elem/ErrorMsg';
+import AddGoalBtn from '../components/common/elem/btn/AddGoalBtn';
 
 import { userId } from '../recoil/userAtoms';
 
@@ -26,7 +25,6 @@ const Home = () => {
   const { id } = useRecoilValue(userId);
   const { isLoading, isError, goals } = useUserGoalsData({ getUserId: Number(id) });
   useBadgesData();
-  const navigate = useNavigate();
 
   return (
     <Wrapper>
@@ -76,22 +74,14 @@ const Home = () => {
             </>
           </>
         )}
-        <AddGoalBtn onClick={() => navigate('/goals/post/type')}>
-          <IconWrapper>
-            <Icon
-              width={20}
-              height={20}
-              color={'gray400'}
-              path='M19.3333 11.3332H11.3333V19.3332H8.66663V11.3332H0.666626V8.6665H8.66663V0.666504H11.3333V8.6665H19.3333V11.3332Z'
-            />
-          </IconWrapper>
-        </AddGoalBtn>
       </ContentWrapper>
+      <AddGoalBtn />
     </Wrapper>
   );
 };
 
 const Wrapper = styled.div`
+  position: relative;
   display: flex;
   flex-direction: column;
   gap: 10px;
@@ -118,31 +108,9 @@ const InfoText = styled.div`
 `;
 
 const ContentWrapper = styled(Wrapper)`
-  padding: 10px;
+  padding: 10px 22px;
   gap: 8px;
   overflow-y: auto;
-`;
-
-const AddGoalBtn = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding: 14px 0;
-  border-radius: 12px;
-  background-color: white;
-  :hover {
-    cursor: pointer;
-  }
-`;
-
-const IconWrapper = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  width: 32px;
-  height: 32px;
 `;
 
 export default Home;
