+++ src/pages/LookupGoals.tsx
@@ -2,8 +2,8 @@ import React, { useRef } from 'react';
 import styled from 'styled-components';
 
 import ImpendingGoals from '../components/goal/lookup/ImpendingGoals';
-
 import GroupGoals from '../components/goal/lookup/GroupGoals';
+import AddGoalBtn from '../components/common/elem/btn/AddGoalBtn';
 
 import RouteChangeTracker from '../shared/RouteChangeTracker';
 
@@ -17,11 +17,13 @@ const LookupGoals = () => {
         <ImpendingGoals />
       </div>
       <GroupGoals />
+      <AddGoalBtn />
     </Wrapper>
   );
 };
 
 const Wrapper = styled.div`
+  position: relative;
   padding-top: 20px;
   display: flex;
   flex-direction: column;
